refactor(PokemonList): changed to use usePokemon and extract main functionality to usePokemon
