+++ src/hooks/usePokemons/usePokemons.tsx
@@ -1,0 +1,32 @@
+import { useState, useEffect } from "react";
+import { API } from "@/constants/apiPaths";
+import { ERROR_MESSAGE } from "@/constants/errorMessage";
+import { Pokemon } from "@/interfaces/pokemon";
+
+export const usePokemons = () => {
+  const [pokemons, setPokemons] = useState<Pokemon[]>([]);
+  const [error, setError] = useState<string | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchPokemons = async () => {
+      try {
+        const url = `${API.BASE_URL}${API.PATHS.GET_ALL_POKEMON}`;
+        const response = await fetch(url);
+
+        if (!response.ok) throw new Error(ERROR_MESSAGE.TRY_AGAIN_LATER);
+
+        const data = await response.json();
+        setPokemons(data);
+      } catch (err) {
+        if (err instanceof Error) setError(err.message);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchPokemons();
+  }, []);
+
+  return { pokemons, error, isLoading };
+};
